+++ app/src/main/java/com/mango/marvelworld/domain/repository/list/CharactersListDataRepository.kt
@@ -3,5 +3,5 @@ package com.mango.marvelworld.domain.repository.list
 import com.mango.marvelworld.domain.models.CharacterDataContainer
 
 interface CharactersListDataRepository {
-    suspend fun getCharacters(offset : Int) : CharacterDataContainer
+    suspend fun getCharacters(offset : Long) : CharacterDataContainer
 }
